AMoRE-I updated
Parallel world for AMoRE-II is maded in AMoRE-I, so I added to option the parallel world installed in AMoRE-II case.

MuonVetoSD is not working with AMoRE-I. This is because that I changed CupVetoSD mother layer number.
Uptonow, it is changed back to original CupVetoSD.
Later I will add to AmoreVetoSD for AMoRE-II.

Macro was cleaned up for AMoRE-I neutron simulation.